requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.51.0
django-braces==1.15.0
django-embed-video==1.4.7
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23.1
urllib3==1.26.12
whitenoise==6.2.0
